requirements.txt
Pillow>=10.3.0
openpyxl>=3.1.5
mysql-connector-python>=9.1.0